+++ GeosCore/fast_jx_mod.F90
@@ -87,6 +87,7 @@ MODULE FAST_JX_MOD
   INTEGER :: id_HOI,    id_IO,   id_OIO,    id_INO,   id_IONO
   INTEGER :: id_IONO2,  id_I2O2, id_CH3I,   id_CH2I2, id_I2O4
   INTEGER :: id_I2O3
+  INTEGER :: id_NIT,    id_NITs, id_SALA,   id_SALC
 
   ! Needed for scaling JNIT/JNITs photolysis to JHNO3
   REAL(fp)      :: JscaleNITs, JscaleNIT, JNITChanA, JNITChanB
@@ -1724,6 +1725,10 @@ CONTAINS
        id_CH2I2    = IND_('CH2I2'   )
        id_I2O4     = IND_('I2O4'    )
        id_I2O3     = IND_('I2O3'    )
+       id_NIT      = IND_('NIT'     )
+       id_NITs     = IND_('NITs'    )
+       id_SALA     = IND_('SALA'    )
+       id_SALC     = IND_('SALC'    )
 
        ! Print info
        IF ( Input_Opt%amIRoot .and. Input_Opt%Verbose ) THEN
@@ -5276,9 +5281,9 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  SUBROUTINE PHOTRATE_ADJ( Input_Opt, State_Diag, State_Met,                 &
-                           I,         J,          L,                         &
-                           FRAC,      RC                                    )
+  SUBROUTINE PHOTRATE_ADJ( Input_Opt, State_Chm, State_Diag,                 &
+                           State_Met, I,         J,                          &
+                           L,         FRAC,      RC                         )
 !
 ! !USES:
 !
@@ -5286,10 +5291,13 @@ CONTAINS
     USE Input_Opt_Mod,  ONLY : OptInput
     USE State_Diag_Mod, ONLY : DgnState
     USE State_Met_Mod,  ONLY : MetState
+    USE State_Chm_Mod,  ONLY : ChmState
+    USE ERROR_MOD,      ONLY : SAFE_DIV
 !
 ! !INPUT PARAMETERS:
 !
     TYPE(OptInput), INTENT(IN)    :: Input_Opt  ! Input_Options object
+    TYPE(ChmState), INTENT(IN)    :: State_Chm  ! Chemistry State object
     TYPE(MetState), INTENT(IN)    :: State_Met  ! Meteorology State object
     INTEGER,        INTENT(IN)    :: I, J, L    ! Lon, lat, lev indices
     REAL(fp),       INTENT(IN)    :: FRAC       ! Result of SO4_PHOTFRAC,
@@ -5303,11 +5311,6 @@ CONTAINS
     INTEGER,        INTENT(OUT)   :: RC         ! Success or failure
 !
 ! !REMARKS:
-!  NOTE: The netCDF diagnostics are attached in DO_FLEXCHEM so that we have
-!  access to the adjusted rates.  Only the bpch diagnostics are updated
-!  here.
-!    -- Bob Yantosca, 19 Dec 2017
-!
 !  %%%% NOTE: WE SHOULD UPDATE THE COMMENTS TO MAKE SURE THAT WE DO      %%%%
 !  %%%% NOT KEEP ANY CONFLICTING OR INCORRECT INFORMATION (bmy, 3/28/16) %%%%
 !
@@ -5319,68 +5322,77 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
 !
-    REAL(fp) :: C_O2,     C_N2, C_H2,   ITEMPK, RO1DplH2O
-    REAL(fp) :: RO1DplH2, RO1D, NUMDEN, TEMP,   C_H2O
+    ! Scalars
+    REAL(fp) :: C_NIT,       C_SALA,     FAC
+    REAL(fp) :: JScale_NITs, JScale_NIT, JNITChanA, JNITChanB
 
-    !=================================================================
+    !========================================================================
     ! PHOTRATE_ADJ begins here!
-    !=================================================================
+    !========================================================================
 
     ! Initialize
-    RC      = GC_SUCCESS
-    TEMP    = State_Met%T(I,J,L)                                 ! K
-    NUMDEN  = State_Met%AIRNUMDEN(I,J,L)                         ! molec/cm3
-    C_H2O   = State_Met%AVGW(I,J,L) * State_Met%AIRNUMDEN(I,J,L) ! molec/cm3
+    ! Also zero NIT and NITs photolysis rxns here (avoids an ELSE block)
+    RC                    = GC_SUCCESS
+    ZPJ(L,RXN_JNITSa,I,J) = 0.0_fp
+    ZPJ(L,RXN_JNITSb,I,J) = 0.0_fp
+    ZPJ(L,RXN_JNITa,I,J)  = 0.0_fp
+    ZPJ(L,RXN_JNITb,I,J)  = 0.0_fp
 
-    ! For all mechanisms. Set the photolysis rate of NITs and NIT to a
-    ! scaled value of JHNO3. NOTE: this is set in geoschem_config.yml
-    IF ( Input_Opt%hvAerNIT ) THEN
+    !========================================================================
+    ! NITRATE PHOTOLYSIS (Viral Shah & Bob Yantsoca, 30 Jan 2023)
+    !
+    ! For all fullchem simulations. Set the photolysis rate of NITs and NIT
+    ! to a scaled value of JHNO3 (scale factors can be specified in the
+    ! in geoschem_config.yml configuration file).
+    !
+    ! Allow particulate nitrate photolysis in the troposphere only.
+    !========================================================================
+    IF ( Input_Opt%hvAerNIT .and. State_Met%InTroposphere(I,J,L) ) THEN
 
-       ! Get the photolysis scalars read in from geoschem_config.yml
+       ! Get NIT and seasalt concentrations [molec cm-3]
+       C_NIT      = State_Chm%Species(id_NIT)%Conc(I,J,L)
+       C_SALA     = State_Chm%Species(id_SALA)%Conc(I,J,L)
+
+       ! Scaling factor for J(NIT)
+       ! Prevent div by zero, and also set to a minimum of 0.1
+       FAC        = SAFE_DIV( C_SALA, ( C_SALA + C_NIT ), 1.0_fp )
+       FAC        = MAX( 0.1_fp, FAC )
+
+       ! Scale factors for NITs and NIT
        JscaleNITs = Input_Opt%hvAerNIT_JNITs
        JscaleNIT  = Input_Opt%hvAerNIT_JNIT
-       ! convert reaction channel % to a fraction
-       JNITChanA  = Input_Opt%JNITChanA
-       JNITChanB  = Input_Opt%JNITChanB
-       JNITChanA  = JNITChanA / 100.0_fp
-       JNITChanB  = JNITChanB / 100.0_fp
+
+       ! Convert reaction channel % to a fraction
+       JNITChanA  = Input_Opt%JNITChanA / 100.0_fp
+       JNITChanB  = Input_Opt%JNITChanB / 100.0_fp
+
        ! Set the photolysis rate of NITs
        ZPJ(L,RXN_JNITSa,I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNITs
        ZPJ(L,RXN_JNITSb,I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNITs
+
        ! Set the photolysis rate of NIT
-       ZPJ(L,RXN_JNITa,I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNIT
-       ZPJ(L,RXN_JNITb,I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNIT
-       ! Adjust to scaling for channels set in geoschem_config.yml
+       ZPJ(L,RXN_JNITa, I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNIT * FAC
+       ZPJ(L,RXN_JNITb, I,J) = ZPJ(L,RXN_JHNO3,I,J) * JscaleNIT * FAC
+
        ! NOTE: channel scaling is 1 in FJX_j2j.dat, then updated here
        ZPJ(L,RXN_JNITSa,I,J) = ZPJ(L,RXN_JNITSa,I,J) * JNITChanA
-       ZPJ(L,RXN_JNITa,I,J) = ZPJ(L,RXN_JNITa,I,J) * JNITChanA
+       ZPJ(L,RXN_JNITa, I,J) = ZPJ(L,RXN_JNITa, I,J) * JNITChanA
        ZPJ(L,RXN_JNITSb,I,J) = ZPJ(L,RXN_JNITSb,I,J) * JNITChanB
-       ZPJ(L,RXN_JNITb,I,J) = ZPJ(L,RXN_JNITb,I,J) * JNITChanB
-
-    ! Gotcha to set JNIT and JNITs to zero if hvAerNIT switch is off
-    ELSE
-
-       ! Set the photolysis rate of NITs to zero
-       ZPJ(L,RXN_JNITSa,I,J) = 0.0_fp
-       ZPJ(L,RXN_JNITSb,I,J) = 0.0_fp
-       ! Set the photolysis rate of NIT to zero
-       ZPJ(L,RXN_JNITa,I,J) = 0.0_fp
-       ZPJ(L,RXN_JNITb,I,J) = 0.0_fp
-
+       ZPJ(L,RXN_JNITb, I,J) = ZPJ(L,RXN_JNITb, I,J) * JNITChanB
     ENDIF
 
-    !==============================================================
+    !========================================================================
     ! SPECIAL TREATMENT FOR H2SO4+hv -> SO2 + 2OH
     !
     ! Only allow photolysis of H2SO4 when gaseous (SDE 04/11/13)
-    !==============================================================
+    !========================================================================
 
     ! Calculate if H2SO4 expected to be gaseous or aqueous
     ! Only allow photolysis above 6 hPa
     ! RXN_H2SO4 specifies SO4 + hv -> SO2 + OH + OH
     ZPJ(L,RXN_H2SO4,I,J) = ZPJ(L,RXN_H2SO4,I,J) * FRAC
 
-    !==============================================================
+    !========================================================================
     ! SPECIAL TREATMENT FOR O3+hv -> O+O2
     !
     ! [O1D]ss=J[O3]/(k[H2O]+k[N2]+k[O2])
@@ -5389,13 +5401,12 @@ CONTAINS
     ! We don't want to do this if strat-chem is in use, as all
     ! the intermediate reactions are included - this would be
     ! double-counting (SDE 04/01/13)
-    !==============================================================
+    !========================================================================
 
     ! Need to subtract O3->O1D from rate
     ! RXN_O3_1 specifies: O3 + hv -> O2 + O
     ! RXN_O3_2 specifies: O3 + hv -> O2 + O(1D)
-    ZPJ(L,RXN_O3_1,I,J) = ZPJ(L,RXN_O3_1,I,J) &
-                        - ZPJ(L,RXN_O3_2,I,J)
+    ZPJ(L,RXN_O3_1,I,J) = ZPJ(L,RXN_O3_1,I,J) - ZPJ(L,RXN_O3_2,I,J)
 
   END SUBROUTINE PHOTRATE_ADJ
 !EOC